Use the Repository has the data container
Given that the repository object is created and managed by
the RepositoryManager, there's no reason to keep data in the
repository manager object. We can keep all the data in the repository
object and have it trigger events for data updates, while events
that are not directly related to data changes in the repository data
model are triggered directly from the repository manager.

This way, the manager manages things and the repository models things,
which feels like a nicer separation of duties.