Fix src/UnityExtension project
To be able to prototype with the src/UnityExtension project directly,
the meta files for the native libraries need to be configured correctly,
and they were only set in the package project.

Move them to the lib folder and copied them to all locations where native
libraries live, so we always have the correctly configured meta file
everywhere.