Remove redundant overload and simplify comparison
Other places in the code call IsFavourite with other types of objects
that aren't strings (like a GitBranch instance). We either have overloads
for all of them or we just take a string and the caller should know
what to pass us, less of an API footprint to worry about.