Fix switching views
Switching views was happening in such a way that the new view would not get
OnDataUpdate called between OnEnable and OnUI, because OnDataUpdate only
gets called during layout events, and we switch the view (calling OnEnable on it)
during the onmouseup event. That meant that the newly-active view would not load
data until the next ui cycle, which causes exceptions to be thrown, given
that the rendering code relies on data to always be there in some way (either
cached or empty or whatever, but *something* needs to exist)

This ensures that OnDataUpdate gets called on the new view manually. This means
that the event type for OnUI is going to be mouseup... which might confuse the
new view if it's trying to handle it? I think it's probably fine because by
this point the event object has been marked as used, which means other controls
that want to handle the event will ignore it.