A serializable Texture2D wrapper to fix icon loading
Fixes #18

Texture2D objects get thrown away by Unity even though they're serialized, because Unity wants to be very conservative about the things it keeps in memory, and games can load large amounts of textures during play mode which Unity clears up. This means that it also throws away all the textures we load, so we lose all the icons.

This adds a class that decomposes Texture2D into separate fields that Unity won't throw away, so we can recreate textures on demand whenever we detect that it has thrown the texture away.