Update Git Status on Commit detected
So there is a bit of a race condition there. We were hoping that when
the index changed event comes through the commit operation is complete.
But that is not the case. We don't reliably get a second index
changed event when it's cleared. So I need to use the commit detection
to say we need to refresh status.