Make it possible to branch out on task success and failure. Cleanups
Kill some unused code in the task system, mainly the Defer paths
which we aren't using anymore
Add a way of calling different tasks depending on the success or failure
of the previous task.
Add some different ways of setting the result of a task beyond the
value returned by the previous task, mainly for seeding initial
values.