Make downloading parallel and fix concurrent threading
Make the downloads in pairs (file and md5) in a downloader
class that can run them all parallel.

I kept getting deadlocks and it's something to do with how the
concurrent scheduler is set up, so replaced it with a scheduler
that just fires a thread per task and does nothing smart about it.