Invalidate caches as soon as we know we need to run commands
Unity might reload the domain before we're finished running commands,
which means any data refresh coming from the file watcher might get lost
and the cache never updated with it. This tries triggering the data
updates by invalidating the cache, so that if Unity reloads we'll
notice that the cache needs refreshing and issue the commands again
as part of the normal cache validation process.