Fix tests that rely on having a git zip around
Since we aren't embedding the zips in the dll anymore, tests that rely on
them were trying to download things, and that causes timeouts. Embed
the zips in the test assembly so they're available for the tests, and
put them in place when starting up the test web server as well, so they
can be served for downloading tests.