Making sure lfs and git merge configs are set up for everyone
Fixes #732

If the project isn't initialized by the extension on the local machine,
the lfs hooks and mergecmd configurations won't be set properly. These
things are as important as the credential helper configuration steps,
so do them at startup.

To make play mode switches faster, cache the git installation state in
the settings and only run the installers once a day, and validate the
existing git/lfs versions once at Unity startup (if we know where they
are from the cached settings)