Validate git installations by version and fix setting up custom gits
Since we know which git and git lfs versions we want and we already
check them in the settings view, switch to using version checks instead
of md5 everywhere. If we find a good enough version, use it.

Make sure we do update checks once a day for git and git lfs.

Go and find lfs if we have a git custom path on all OSes

Make sure we restart the repository manager and everything else
that relies on git when we switch gits, since outputs may change.