Make the repo dir -> project dir mapping in IO operations more transparent
Every time we do an IO operation, we need to convert any relative paths we
have to be based off of the real working directory. That's easier to do with
an NPath object than with a string. This introduces a wrapper object that
does the conversion and passes the call onto IFileSystem, so that NPath
itself doesn't know about it, so it's easier to audit all calls and not miss
any conversions.

Context:
NPath/FileSystem are working under the assumption that the working directory
of the process is the repo root. This is a lie - Unity sets the working dir
to the project path when it loads a project. We lie so it's easier
to spawn our processes from the root of the git repository and to show all
changes, even though they're outside the Unity project path.