configure: UnityYamlMerge as a merge tool
UnityYamlMerge is a merge tool which can help with conflict resolution.
However, currently we configure it using the merge section instead of
the mergetool section, meaning that it's not being used.  In addition,
we configure it in .gitattributes, even though that has no effect.

Configure this tool in the correct section of the Git config file and
skip inserting it into the gitattributes file.