Fix the git log and git status output processors. Split status field into two.
The git status output processor was not handling all the variations
that the command outputs for the status flags. Make it actually follow
what the manual says.

In the process, the status field of a git status entry has now been
split into indexStatus and workTreeStatus fields, to provide all the
underlying information that the command provides.

Staged and Unmerged can be obtained via new properties in the GitStatusEntry
structure, or calculated directly from the indexStatus and workTreeStatus fields.